Refs #14: Authentication opcode: unit test for JWT
